# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_readygeneration_readygeneration-backend_pkg_response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.APIError'
      meta:
        $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_readygeneration_readygeneration-backend_pkg_response.Meta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handler.createOrgRequest:
    properties:
      city:
        type: string
      ein:
        type: string
      mission:
        type: string
      name:
        type: string
      org_type:
        type: string
      phone:
        type: string
      slug:
        type: string
      state:
        type: string
      website:
        type: string
      zip:
        type: string
    required:
    - name
    - org_type
    - slug
    type: object
...
  internal_handler.updateOrgRequest:
    properties:
      city:
        type: string
      logo_url:
        type: string
      mission:
        type: string
      name:
        type: string
      phone:
        type: string
      state:
        type: string
      website:
        type: string
      zip:
        type: string
    type: object
  internal_handler.upsertProfileRequest:
    properties:
      annual_budget:
        type: integer
      focus_issues:
        items:
          type: string
        type: array
      has_501c3:
        type: boolean
      has_audited_financials:
        type: boolean
      has_indirect_cost_rate:
        type: boolean
      indirect_cost_rate_pct:
        type: number
      narrative:
        type: string
      num_employees:
        type: integer
      num_volunteers:
        type: integer
      populations_served:
        items:
          type: string
        type: array
      prior_federal_grants:
        type: boolean
      program_areas:
        items:
          type: string
        type: array
      service_areas:
        items:
          type: string
        type: array
      years_operating:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Grant intelligence platform API
  title: ReadyGeneration API
  version: "1.0"
paths:
...
  /orgs:
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createOrgRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - organizations
  /orgs/{id}:
    get:
      parameters:
      - description: Organization UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get an organization by ID
      tags:
      - organizations
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateOrgRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update an organization
      tags:
      - organizations
  /orgs/{id}/profile:
    get:
      parameters:
      - description: Organization UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get the detailed profile for an organization
      tags:
      - organizations
    put:
      consumes:
      - application/json
      parameters:
      - description: Organization UUID
        in: path
        name: id
        required: true
        type: string
      - description: Profile data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.upsertProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create or update an organization's eligibility profile
      tags:
      - organizations
...
  /orgs/{org_id}/grants/{grant_id}/score:
    get:
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      - description: Grant UUID
        in: path
        name: grant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Get stored compatibility score
      tags:
      - scoring
    post:
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      - description: Grant UUID
        in: path
        name: grant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Compute compatibility score for an org/grant pair
      tags:
      - scoring
  /orgs/{org_id}/score-all:
    post:
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Trigger bulk scoring of all active grants for an org
      tags:
      - scoring
  /orgs/{org_id}/top-grants:
    get:
      parameters:
      - description: Organization UUID
        in: path
        name: org_id
        required: true
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_readygeneration_readygeneration-backend_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List top-scored grants for an organization
      tags:
      - scoring
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"